rename images too to snake_case
