+++ src/renderer/pages/console_mirror/new_connection_list/new_connection_item.tsx
@@ -3,7 +3,7 @@ import AddIcon from "@mui/icons-material/Add";
 import IconButton from "@mui/material/IconButton";
 import Tooltip from "@mui/material/Tooltip";
 
-import { ReactComponent as WiiIcon } from "@/styles/images/wii-icon.svg";
+import { ReactComponent as WiiIcon } from "@/styles/images/wii_icon.svg";
 
 type NewConnectionItemProps = {
   nickname?: string;
