+++ src/renderer/pages/console_mirror/saved_connections_list/saved_connection_item.tsx
@@ -18,7 +18,7 @@ import { ExternalLink as A } from "@/components/external_link";
 import { LabelledText } from "@/components/labelled_text";
 import { useToasts } from "@/lib/hooks/use_toasts";
 import { useServices } from "@/services";
-import { ReactComponent as WiiIcon } from "@/styles/images/wii-icon.svg";
+import { ReactComponent as WiiIcon } from "@/styles/images/wii_icon.svg";
 
 const path = window.electron.path;
